testes

+++ bitcommerce/src/app/page.tsx
@@ -1,37 +1,31 @@
 'use client'
+
 import { Container, Grid2, Typography } from "@mui/material";
 import Header from "./components/Header";
 import Footer from "./components/Footer";
 import { Cart, IProduto } from "./components/cart";
 import { ProductList } from "./components/cart/ProductList";
-import { useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 
 export default function Home() {
 
-  const [cart, setCart] = useState<IProduto[]>(() => {
-    if (typeof window !== "undefined") {
-      const storedCart = localStorage.getItem("cart");
+  const [cart, setCart] = useState<IProduto[]>([]);
 
-      try {
-        return storedCart ? JSON.parse(storedCart) : [];
-      } catch (error) {
-        console.error(error);
-        return [];
+  useEffect(() => {
+      if (typeof window !== "undefined") {
+        const storedCart = localStorage.getItem("cart");
+  
+        try {
+          setCart(storedCart ? JSON.parse(storedCart) : []);
+        } catch (error) {
+          console.error(error);
+          // return [];
+          setCart([]);
+        }
       }
-    }
-  });
+      // return [];
 
-  // useEffect(() => {
-    // if (storedCart) {
-    //   try {
-    //     setCart(JSON.parse(storedCart));
-        
-    //   } catch (error) {
-    //     console.error(error);
-    //     setCart([]);
-    //   }
-    // }
-  // }, [])
+  }, [])
 
   useEffect(() => {
     localStorage.setItem("cart", JSON.stringify(cart));
@@ -40,22 +34,45 @@ export default function Home() {
 
   const addToCart = (produto: IProduto) => {
     setCart((prevCart) => {
-      const newCart = [...prevCart, produto];
-      localStorage.setItem("cart", JSON.stringify(newCart));
-      return newCart;
+      const existingProductIndex = prevCart.findIndex((item) => item.id === produto.id);
+
+      if (existingProductIndex >=0) {
+        const updatedCart = [...prevCart];
+        updatedCart[existingProductIndex] = {
+          ...updatedCart[existingProductIndex],
+          quantidade: updatedCart[existingProductIndex].quantidade + 1,
+        };
+        localStorage.setItem("cart", JSON.stringify(updatedCart));
+        
+        return updatedCart;
+
+      } else {
+        const newCart = [...prevCart,{ ...produto, quantidade: 1 }];
+        localStorage.setItem("cart", JSON.stringify(newCart));
+
+        return newCart;
+      }
     });
   };
 
   const  removeFromCart = (index: number) => {
     setCart((prevCart) => {
-      const newCart = prevCart.filter((_, i) => i !== index);
-      localStorage.setItem("cart", JSON.stringify(newCart));
-      return newCart
+     const updatedCart = [...prevCart];
+     if (updatedCart[index].quantidade > 1) {
+      updatedCart[index] = {
+        ...updatedCart[index],
+        quantidade: updatedCart[index].quantidade -1,
+      }
+     } else {
+      updatedCart.splice(index, 1);
+     }
+     localStorage.setItem("cart", JSON.stringify(updatedCart));
+     return updatedCart;
     });
   }
 
   return (
-    <>
+    <div>
       <Header />
       <Container>
         <Typography variant="h3" component="h1" align="center" gutterBottom>
@@ -63,9 +80,7 @@ export default function Home() {
         </Typography>
         <Grid2 container spacing={3} sx={{ display: "flex", flexWrap: "wrap"}}>
           <Grid2 sx={{ display: "flex", flexWrap: "wrap"}}>
-            {/* <Grid2> */}
               <ProductList addToCart={addToCart} />
-            {/* </Grid2> */}
           </Grid2>
           <Grid2>
             <Cart cart={cart} removeFromCart={removeFromCart} />
@@ -73,6 +88,6 @@ export default function Home() {
         </Grid2>
       </Container> 
       <Footer />
-    </>
+    </div>
   );
 }
